Move namedtuple to compatibility
